Display total item count

+++ model/category.go
@@ -3,5 +3,5 @@ package model
 type Category struct {
 	Description string
 	Items       []Item
-	ItemCount   int
+	ItemCount   int64
 }
